Add sensitivity note to Module 2b menstrual period question
Added note before the question about last menstrual period:
'Note: we are going to ask some questions which are sensitive in
some cultures and regions. If these answers are too sensitive for
you, please just write the answers down on a piece of paper
separately and keep it with you, ready for you to give the
information to your doctor when they ask.'

This change applies to all three domains (UK/US/AU).